feat: udistrital/sga_documentacion#24 Refactorización de micro-mids migración de logica de controladores.
